\docType{package}
\name{lookfor-package}
\alias{lookfor-package}
\title{Find What You're Looking For}
\description{lookfor is an R port of Stata's \code{lookfor} command, which helps you find stuff you're looking for inside a dataset (e.g., variables, variable values, variable labels, etc.).}
\seealso{\code{\link{lookfor}}}
