\docType{package}
\name{lookfor}
\alias{lookfor}
\alias{lookfor-package}
\title{Find what you're looking for}
\description{lookfor is an R port of Stata's \code{lookfor} command, which helps you find stuff you're looking for inside a dataset (e.g., variables, variable values, variable labels, etc.).}
